Update the gem file structure and files to modern standard. No functional code changes.
These commits update all files and the file system hierarchy to the current
modern standard for gems. This should not change anything about the
functionality of the code.

Rspec is updated to work around deprecations since this gem was last updated.

A LICENSE.md file was added with the appropriate license text.

The new binaries that are installed for convenience with new gems
are present to make dev a bit easier.

Using HTTPS for rubygems.

lib/srp.rb is the main file now, not srp-rb.rb which was not
correct for this gem since it was to be required with
require srp.

The gemspec was updated to the current default standard and all
appropriate info carried over.

Added starter .rubocop.yml and .travis.yml config files.