Resolves major issue where generated hex strings were sometimes of invalid odd length.
Oftentimes large BigNum will result in odd length hex strings
when derived with:

  format('%x', n)

This is somewhat less so (why?) when done with:

  n.to_s(16) # upcased but often more even length

To resolve this, created a custom method to normalize the
conversion of large numbers to hex by checking length,
and if odd pre-pending with a '0' string.

Also downcasing all hex strings.

Without this *many* clients and Ruby use of the generated
hex would fail with errors due to inability to parse an
odd length hex string.